pyproject.toml
[tool.poetry]
name = "psiagent"
version = "2.0.0"
description = "PsiAgent - Assistente Clínico Local. A privacy-focused clinical assistant with AI-powered tools for psychologists."
authors = ["Luiz Lopes <luizlopesrandom@gmail.com>"]
license = "MIT"
readme = "README.md"
keywords = ["clinical", "psychiatry", "psychology", "ai", "langgraph", "fastapi"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Healthcare Industry Opensource",
    "Topic :: Software Development :: Libraries",
    "Environment :: Web Environment",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: MIT License",
]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11,<3.14"

# Servidor Web & API
fastapi = "^0.109.0"
uvicorn = {version = "^0.27.0", extras = ["standard"]}
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"

# Banco de Dados
sqlalchemy = "^2.0"
psycopg2-binary = "^2.9"

# Agente & IA
langchain = "^0.2.0"
langchain-core = "^0.2.20"
langchain-ollama = "^0.1.0"
langgraph = "^0.1.0"

# Protocolo MCP (Tools)
mcp = "^1.0.0"
sse-starlette = "^2.0.0"

# Utilitários
pydantic = "^2.0"
requests = "^2.31.0"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.9.0"
ruff = "^0.0.290"
mypy = "^1.5.0"
ipython = "^8.15.0"
ipdb = "^0.13.0"

[tool.poetry.scripts]
psiagent-serve = "main:run_server"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'
exclude = '''/(\.git|\.hg|\.mypy_cache|\.tox|\.venv|_build|buck-out|build|dist|__pycache__)/'''

[tool.ruff]
line-length = 100
target-version = "py312"
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "C",      # flake8-comprehensions
    "B",      # flake8-bugbear
]
ignore = [
    "E501",   # line too long, handled by black
    "C901",   # too complex
]

[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
disallow_untyped_defs = false
ignore_missing_imports = true
strict_optional = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
asyncio_mode = "auto"

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true
